refactor and write docs
