+++ alchemy/src/cloudflare/index.ts
@@ -5,6 +5,7 @@ export * from "./ai.ts";
 export * from "./analytics-engine.ts";
 export * from "./api-error.ts";
 export * from "./api-gateway-operation.ts";
+export * from "./api-schema.ts";
 export * from "./api-shield.ts";
 export * from "./api.ts";
 export * from "./assets.ts";
@@ -43,7 +44,6 @@ export * from "./react-router.ts";
 export * from "./redirect-rule.ts";
 export * from "./redwood.ts";
 export * from "./route.ts";
-export * from "./schema.ts";
 export * from "./secret-key.ts";
 export * from "./secret.ts";
 export * from "./secrets-store.ts";
